Implemented Wild Charge
 On branch feral
 Changes to be committed:
	deleted:    sim/druid/_feral_charge.go
	modified:   sim/druid/feral/TestFeral.results
	modified:   sim/druid/feral/feral_test.go
	modified:   sim/druid/feral/rotation.go
	modified:   sim/druid/talents.go
